refactor(web): #63 프로필 페이지 LoginButton·UserInfo 컴포넌트 분리
