+++ apps/web/src/app/profile/page.tsx
@@ -1,42 +1,11 @@
 "use client";
 
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { useRouter } from "next/navigation";
 import Link from "next/link";
 import { useAuth } from "@/contexts/AuthContext";
-
-type ProviderLabel = "google" | "kakao" | "naver";
-
-const PROVIDER_LABELS: Record<ProviderLabel, string> = {
-  google: "Google",
-  kakao: "카카오",
-  naver: "네이버",
-};
-
-function formatDate(dateString: string): string {
-  const date = new Date(dateString);
-  return date.toLocaleDateString("ko-KR", {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-  });
-}
-
-function formatDateTime(dateString: string): string {
-  const date = new Date(dateString);
-  return date.toLocaleDateString("ko-KR", {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-    hour: "2-digit",
-    minute: "2-digit",
-  });
-}
-
-function getProviderLabel(provider: string | null): string {
-  if (!provider) return "알 수 없음";
-  return PROVIDER_LABELS[provider as ProviderLabel] ?? provider;
-}
+import { LoginButton } from "@/components/auth/LoginButton";
+import { UserInfo } from "@/components/auth/UserInfo";
 
 const MAX_REGIONS = 20;
 
@@ -47,6 +16,10 @@ function InterestRegionsSection({ regions }: { regions: string[] }) {
   const [isSaving, setIsSaving] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const [justSaved, setJustSaved] = useState(false);
+
+  useEffect(() => {
+    setDraft(regions);
+  }, [regions]);
 
   const isDirty =
     draft.length !== regions.length ||
@@ -95,7 +68,7 @@ function InterestRegionsSection({ regions }: { regions: string[] }) {
         관심 지역을 등록하면 실거래가, 청약 공고 알림에 활용됩니다.
       </p>
 
-      <div className="mb-4 flex gap-2">
+      <div className="mb-4 flex flex-col gap-2 sm:flex-row">
         <input
           type="text"
           value={input}
@@ -107,7 +80,7 @@ function InterestRegionsSection({ regions }: { regions: string[] }) {
             }
           }}
           placeholder="예: 서울 강남구"
-          className="flex-1 rounded-lg border px-3 py-2 text-sm"
+          className="min-w-0 flex-1 rounded-lg border px-3 py-2 text-sm"
           aria-label="관심 지역 입력"
         />
         <button
@@ -202,12 +175,7 @@ export default function ProfilePage() {
           <p className="mb-6 text-sm text-muted-foreground">
             프로필을 보려면 먼저 로그인해주세요.
           </p>
-          <Link
-            href="/login"
-            className="rounded-lg bg-primary px-6 py-3 text-sm font-medium text-primary-foreground hover:bg-primary/90"
-          >
-            로그인하기
-          </Link>
+          <LoginButton />
         </main>
       </div>
     );
@@ -240,63 +208,10 @@ export default function ProfilePage() {
       <main className="mx-auto max-w-lg px-4 py-12">
         <h1 className="mb-8 text-2xl font-bold">내 프로필</h1>
 
-        <div className="rounded-lg border bg-card p-6">
-          {/* Profile avatar placeholder */}
-          <div className="mb-6 flex items-center gap-4">
-            <div className="flex h-14 w-14 items-center justify-center rounded-full bg-primary/10 text-xl font-bold text-primary">
-              {user.nickname?.charAt(0) ?? user.email?.charAt(0) ?? "?"}
-            </div>
-            <div>
-              <p className="font-semibold">
-                {user.nickname ?? "닉네임 없음"}
-              </p>
-              <p className="text-sm text-muted-foreground">
-                {user.email ?? "이메일 없음"}
-              </p>
-            </div>
-          </div>
+        <UserInfo user={user} />
 
-          <hr className="mb-6" />
-
-          {/* Profile details */}
-          <dl className="space-y-4">
-            <div className="flex justify-between">
-              <dt className="text-sm text-muted-foreground">이메일</dt>
-              <dd className="text-sm font-medium">
-                {user.email ?? "등록되지 않음"}
-              </dd>
-            </div>
-            <div className="flex justify-between">
-              <dt className="text-sm text-muted-foreground">닉네임</dt>
-              <dd className="text-sm font-medium">
-                {user.nickname ?? "등록되지 않음"}
-              </dd>
-            </div>
-            <div className="flex justify-between">
-              <dt className="text-sm text-muted-foreground">로그인 방식</dt>
-              <dd className="text-sm font-medium">
-                {getProviderLabel(user.provider)}
-              </dd>
-            </div>
-            <div className="flex justify-between">
-              <dt className="text-sm text-muted-foreground">가입일</dt>
-              <dd className="text-sm font-medium">
-                {formatDate(user.createdAt)}
-              </dd>
-            </div>
-            <div className="flex justify-between">
-              <dt className="text-sm text-muted-foreground">마지막 활동</dt>
-              <dd className="text-sm font-medium">
-                {formatDateTime(user.lastActiveAt)}
-              </dd>
-            </div>
-          </dl>
-        </div>
-
-        {/* Interest regions */}
         <InterestRegionsSection regions={user.interestRegions ?? []} />
 
-        {/* Actions */}
         <div className="mt-6 flex flex-col gap-3">
           <Link
             href="/"
